enhance: remove deprecated code within channel manager
